international-student-hub/backend/requirements.txt
# =============================================================================
# International Student Hub — Taiwan
# requirements.txt — All Python dependencies with pinned versions
# =============================================================================
# Install: pip install -r requirements.txt
# =============================================================================

# --------------------------------------------------------------------------
# Django Core
# --------------------------------------------------------------------------
Django==4.2.13
django-environ==0.11.2          # Reads .env files + env() helper for settings

# --------------------------------------------------------------------------
# Django REST Framework
# --------------------------------------------------------------------------
djangorestframework==3.15.1
django-filter==24.2             # ?filter= query params in DRF views

# --------------------------------------------------------------------------
# Authentication — JWT
# --------------------------------------------------------------------------
djangorestframework-simplejwt==5.3.1

# --------------------------------------------------------------------------
# Database — PostgreSQL
# --------------------------------------------------------------------------
psycopg2-binary==2.9.9          # PostgreSQL driver (binary = no build deps needed)

# --------------------------------------------------------------------------
# CORS — Allow React frontend to call the API
# --------------------------------------------------------------------------
django-cors-headers==4.3.1

# --------------------------------------------------------------------------
# Cache & Task Queue
# --------------------------------------------------------------------------
redis==5.0.4                    # Python Redis client
django-redis==5.4.0             # Django cache backend using Redis
celery==5.3.6                   # Distributed task queue
kombu>=5.3.4                    # Celery's messaging library (pinned for compatibility)

# --------------------------------------------------------------------------
# Static Files (Production)
# --------------------------------------------------------------------------
whitenoise==6.7.0               # Serve static files without a CDN

# --------------------------------------------------------------------------
# LINE Bot
# --------------------------------------------------------------------------
requests==2.31.0                # HTTP client for LINE API calls

# --------------------------------------------------------------------------
# Image handling (for avatar uploads)
# --------------------------------------------------------------------------
Pillow==10.3.0

# --------------------------------------------------------------------------
# Production Server
# --------------------------------------------------------------------------
gunicorn==21.2.0               # WSGI server (used in entrypoint.sh)

# --------------------------------------------------------------------------
# Development Utilities
# --------------------------------------------------------------------------
ipython==8.24.0                 # Better Django shell (python manage.py shell)